Implement Audio Session Capture Source Stub
Add AudioCaptureSource and related method stubs as well as functions
to wrap those methods so that they can be used by obs_source_info.

This structure is based on the WASAPISource plugin included with OBS by
default. Given the vaguely similar scope, and the known necessity of
implementation in C++ rather than C, it made sense to create the source
this way rather than the C-like method used in other plugins.