Implement method to get every active audio session
Hooking an arbitrary process probably requires knowing every process
with an active audio session. Although the goal is to hook a given
audio client, the closest we can get to knowing which processes have
an audio client given documented Windows API is to iterate over every
audio session, which may contain multiple streams with each stream
using a client.

This time, we draw inspiration from win-capture as well as win-wasapi.

win-capture uses an obfuscate function to work around certain
anti-virus programs. Unfortunately, casting void* to function pointers
is explicitly forbidden in C++, rather than a warning as in C. This
means that the obfuscate implementation used for win-capture cannot be
used here. For the time being, the relevant functions are called
directly, but it would be suggested to implement an obfuscator in the
future.