Fix StringFromLPWSTR Length
When given an empty string, StringFromLPWSTR would return a string with
a length of one. This is because the string was resized to accomodate a
null-terminator that would be added by os_wcs_to_utf8. Instead of doing
some fancy trickery, the copied string is just resized to be one less
than the original length with the null-terminator.