Replace snake_case with camelCase
I had been using a bad mix of PascalCase and snake_case. Since style
typically dictates use of snakeCase in C++, instances of snake_case
that can be replaced with camelCase have been. Calls to OBS are still
snake_case and nothing can be done about that without an amount of
effort and redundancy that is not worth pursuing.

This was totally just an excuse to avoid the more important work.