Implement PluginHelper.cmake
ObsPluginHelpers.cmake is included with LibObs, but visual inspection
leads me to believe it's not going to work. The needed functionality
from ObsPluginHelpers has been implemented in PluginHelper.

Important to note that install_plugin_bin_to_data does not include the
pdb file. It probably should, but the debug files probably wouldn't be
helpful either because of the nature of the expected data.

Since PluginHelper isn't really external, it was put into the cmake
directory. FindLibObs has also been placed there just to keep tidy.