Split AudioCaptureSource implementation
While working on the project, it became clear that the combined
declaration and implementation of AudioCaptureSource was becoming
unwieldy, making it difficult to implement new features without a mess
of extern declarations. We aim to resolve this by doing as C++ style
suggests and splitting the AudioCaptureSource into separate header and
implementation files. This split implementation style will attempt to
be enforced in future features when possible.