Implement get-audio-offsets
To hook the target application, we need to know the addresses of the
functions of interest. Currently, it is only expected that we will need
to know the addresses for the GetBuffer and ReleaseBuffer methods in
IAudioRenderClient.

Modeled after (stolen from?) win-capture, get-audio-offsets is
implemented as a standalone project that produces a binary to be run by
the plugin on initialization. Without this separation, we would only be
able to know the addresses for applications using the same architecture
as the plugin (currently 64-bit.) This way, we can produce a 32-bit
binary that can receive the 32-bit addresses. That said, I have no idea
how to do that because even my build of obs-studio only contains the
64-bit binaries for win-capture.

The audio-hook-info header is currently barebones, containing only the
struct for the method addresses. The win-capture equivalent contains
much more information, particularly macros to share with the hook and
the plugin. These were left out as to not be overwhelming and confusing
since they would not be utilized for quite a while.

The copyright here is particularly questionable. The license is still
GPLv2, but this is so heavily based on win-capture that it probably
falls more under Jim's copyright than my own.