Implement ProcessPipe
In libobs, there is a util/pipe.h header for creating process pipes.
Possibly because it is only used in two places, it is undocumented and
not guarded with extern "C". Unfortunately for us, one of the two places
it is used is in win-capture because it is used to load the graphics
offsets from get-graphics-offsets. Since this functionality is useful
and more convenient that having a large blob of handles everywhere,
ProcessPipe was born.

ProcessPipe rolls up the functionality of a child process and redirects
the output for our convenience. It's a marginally nicer wrapper than
util/pipe.h but not by much.

process-pipe was originally called pipe-helper, which makes more sense
to place in the helpers folder but also didn't fully describe its goals.
It's entirely possible that we might see a pipe-helper in the future
that is concerned with generic pipe use and not specifically for child
processes. If that happens, renaming can be reconsidered.