Implement Preinitialization
When the win-capture module is loaded, it spins a new thread to run
load_graphics_offsets, which in turn runs the get-graphics-offsets
programs and stores the results. This is likely done for some amount of
asynchronicity so that loading isn't help up by this process, though I
imagine in practice that it would rarely be an issue. Nonetheless, it
makes sense to do, and so we do. Here, we call it Preinitialization.

At the moment, Preinitialize just does as win-capture's init_hooks does
by running its offset fetch program on a new thread. Unlike win-capture
though, we can add additional preinitialization steps to Preinitialize
rather than into the module load.

This continues with the attempted parity of win-capture.