Restructure get-audio-offsets
The original get-graphics-offsets was particularly beefy, having to
manage with multiple different interfaces. Since we presumably only
have one to deal with, we can simplify some of the structure.

This could be a single file, but the initialization required is a bit
messy, so we retain the multi-file structure and just do some tidying.

In an attempt to avoid "UNIX-Style relative paths", we now include
parent and sibling directories as needed, starting with this target.
This allows us to move audio-hook-info.hpp into the audio-hook directory
where it belongs.