Move and Rename generated files
Files configured by CMake were originally placed back in-source, which
produces clutter. Now, the destination of configured files is relative
to CMAKE_CURRENT_BINARY_DIR, which is the default behavior of
configure_file when given a relative path destination. This also allows
us to name the configured files without the "generated" suffix, as they
are now within an ignored directory.

This commit continues with adjusting to the new audio-hook-info location