Fix incorrect sizeof and Remove dead code
ProcessPipe used the sizeof operator, but I apparently forgot how it's
used. Parentheses are required for types, but not for expressions. I had
somehow managed to do the exact opposite. It seems to have worked, but
it's not exactly safe to keep.

Since we're working in process-pipe, there was a method that never got
implemented. Since we don't need it, it has been removed.